UI WIP
main.py, updated to just include the potential pages we need for now.
cycler_logs.py, updated the log window to be themed the same as system_logs.
live_view.py, updated to demo more of a live view than static data, this includes a 100 rolling sample of test data.